Rename modules.
