+++ src/_gettsim/taxes/einkommensteuer/group_by_ids.py
