+++ src/_gettsim/transfers/arbeitslosengeld_2/group_by_ids.py
